+++ libs/cmcd/src/CmcdEvent.ts
@@ -2,17 +2,15 @@ import type { CmcdEventType } from './CmcdEventType.ts'
 import type { CmcdRequest } from './CmcdRequest.ts'
 
 /**
- * CMCD v2 - Event Mode.
+ * Common Media Client Data (CMCD) version 2 - Event Mode.
  *
- * Represents the event and keys for CMCD v2.
- *
- * This type is used to structure the data for reporting events according to the
- * Common Media Client Data (CMCD) version 2 specification. It encapsulates
- * the reporting event token.
+ * Extends {@link CmcdRequest} with event-specific keys for reporting events
+ * according to the CMCD version 2 specification.
  *
  * @public
  */
 export type CmcdEvent = CmcdRequest & {
+
 	/**
 	 * Custom Event Name
 	 *
@@ -30,27 +28,27 @@ export type CmcdEvent = CmcdRequest & {
 	 * This key MUST only be used in Event mode and MUST be present on all reports. The minimum recommended set of supported events
 	 * are: `ps`, `e`, `t`, and `rr`.
 	 *
-	 * - `abs` - ad break start: The start of an ad break or ad pod which would contain 1 or more sequential ads.
-	 * - `abe` - ad break end: This signals the end of an ad break or ad pod.
-	 * - `ae` - ad end: This token should be used at the end of the current playing ad but before exiting the ad.
-	 * - `as` - ad start: This token should be used when a new ad begins playing within an ad break pod.
-	 * - `b` - The player has entered backgrounded mode if this event is accompanied by the 'bg' key and exited backgrounded mode if not.
-	 * - `bc` - The bitrate being requested by the player, for any object type, has changed.
-	 * - `c` - content ID has changed.
-	 * - `ce` - custom event.
-	 * - `e` - the player has experienced an error. This token MUST be accompanied by a 'ec' key defining the player error code.
-	 * - `h` - hostname has changed.
-	 * - `m` - mute. The user activated the mute control or set the volume to zero.
-	 * - `pc` - playerCollapse: the user activated a control to reduce the player to a smaller size.
-	 * - `pe` - playerExpand: The user activated a control to extend the player to a larger size.
-	 * - `pr` - playback rate change. This event only triggers while the state is playing ('p').
-	 * - `ps` - play state change. This token MUST be accompanied by a 'sta' key carrying the new state.
-	 * - `rr` - response received: This signals the receipt of a response. This event SHOULD be accompanied with the url key.
-	 * - `sk` - skip: the user activated a control to skip an advertisement.
-	 * - `t` - time interval. A default interval of 30 seconds SHOULD be used. This event MUST be supported by all players that support Event mode.
-	 * - `um` - unmute. The user deactivated the mute control or raised the volume above zero if it was previously set to zero.
+	 * - `abs` - ad break start
+	 * - `abe` - ad break end
+	 * - `ae` - ad end
+	 * - `as` - ad start
+	 * - `b` - backgrounded mode
+	 * - `bc` - bitrate change
+	 * - `c` - content ID changed
+	 * - `ce` - custom event
+	 * - `e` - error
+	 * - `h` - hostname changed
+	 * - `m` - mute
+	 * - `pc` - player collapse
+	 * - `pe` - player expand
+	 * - `pr` - playback rate change
+	 * - `ps` - play state change
+	 * - `rr` - response received
+	 * - `sk` - skip
+	 * - `t` - time interval
+	 * - `um` - unmute
 	 *
-	 * Token - one of [abs,abe,ae,as,b,bc,c,ce,e,h,m,pc,pe,ps,rr,sk,t,um]
+	 * Token
 	 */
 	e?: CmcdEventType;
 
@@ -62,16 +60,6 @@ export type CmcdEvent = CmcdRequest & {
 	 * String
 	 */
 	h?: string;
-
-	/**
-	 * Sequence Number
-	 *
-	 * A monotonically increasing integer to identify the sequence of a CMCD report to a target within a session. This MUST be reset to
-	 * zero on the start of a new session-id. Sequence numbers increase independently per each combination of mode and target.
-	 *
-	 * Integer
-	 */
-	sn?: number;
 
 	/**
 	 * Timestamp