+++ libs/cmcd/src/CmcdRequestData.ts
@@ -1,9 +1,0 @@
-import type { Cmcd } from './Cmcd.ts'
-import type { CmcdRequest } from './CmcdRequest.ts'
-
-/**
- * CMCD request data.
- *
- * @public
- */
-export type CmcdRequestData = Cmcd & CmcdRequest