+++ libs/cta/src/CmValue.ts
@@ -1,5 +1,4 @@
 import type { SfItem, SfToken } from '@svta/cml-structured-field-values'
-import type { ValueOrArray } from '@svta/cml-utils'
 import type { CmObjectType } from './CmObjectType.ts'
 import type { CmStreamingFormat } from './CmStreamingFormat.ts'
 import type { CmStreamType } from './CmStreamType.ts'
@@ -13,7 +12,6 @@ export type CmValue =
 	| CmObjectType
 	| CmStreamingFormat
 	| CmStreamType
-	| ValueOrArray<string | SfItem<string>>
 	| string
 	| string[]
 	| number
@@ -22,5 +20,4 @@ export type CmValue =
 	| symbol
 	| SfToken
 	| SfItem
-	| SfItem[]
-	| (number | SfItem)[];
+	| SfItem[];